+++ NNP-NET/LayoutMethods/NNPNET.cpp
@@ -32,11 +32,11 @@ void NNPNet::NNPNET::run(Graph<float>& g, Graph<double>* GT, int pivots)
 
 float* NNPNet::NNPNET::createPivotEmbedding(Graph<float>& g, int pivots)
 {
-	float* embedding = (float*)malloc(g.nodeCount * pivots * sizeof(float));
-	int n = g.nodeCount;
-	int pivot = 0;
+	float* embedding = (float*)malloc(((size_t)g.nodeCount) * ((size_t)pivots) * sizeof(float));
+	size_t n = g.nodeCount;
+	size_t pivot = 0;
 	float* lowest = (float*)malloc(sizeof(float) * n);
-	for (int i = 0; i < pivots; i++) {
+	for (size_t i = 0; i < pivots; i++) {
 		g.getDistances(pivot, embedding + i*n);
 
 		if (i + 1 < pivots) {
@@ -51,7 +51,7 @@ float* NNPNet::NNPNET::createPivotEmbedding(Graph<float>& g, int pivots)
 				}
 				continue;
 			}
-			int p_i = i * n;
+			size_t p_i = i * n;
 			for (int j = 0; j < n; j++) {
 				if (embedding[p_i] < lowest[j]) {
 					lowest[j] = embedding[p_i];
@@ -65,16 +65,16 @@ float* NNPNet::NNPNET::createPivotEmbedding(Graph<float>& g, int pivots)
 		}
 	}
 	float biggest = 0;
-	for (int i = 0; i < g.nodeCount; i++) {
-		for (int j = 0; j < pivots; j++) {
-			if (embedding[j * g.nodeCount + i] > biggest) {
-				biggest = embedding[j * g.nodeCount + i];
+	for (size_t i = 0; i < g.nodeCount; i++) {
+		for (size_t j = 0; j < pivots; j++) {
+			if (embedding[j * ((size_t)g.nodeCount) + i] > biggest) {
+				biggest = embedding[j * ((size_t)g.nodeCount) + i];
 			}
 		}
 	}
-	for (int i = 0; i < g.nodeCount; i++) {
-		for (int j = 0; j < pivots; j++) {
-			embedding[j * g.nodeCount + i] /= biggest;
+	for (size_t i = 0; i < g.nodeCount; i++) {
+		for (size_t j = 0; j < pivots; j++) {
+			embedding[j * ((size_t)g.nodeCount) + i] /= biggest;
 		}
 	}
 	
@@ -83,10 +83,10 @@ float* NNPNet::NNPNET::createPivotEmbedding(Graph<float>& g, int pivots)
 
 float* NNPNet::NNPNET::createPMDSEmbedding(Graph<float>& g, int& dimensions)
 {
-	float* embedding = (float*)malloc(g.nodeCount * dimensions * sizeof(float));
-	int n = g.nodeCount;
+	float* embedding = (float*)malloc(((size_t)g.nodeCount) * ((size_t)dimensions) * sizeof(float));
+	size_t n = g.nodeCount;
 	float* backupY = g.Y;
-	int backupDim = g.outputDim;
+	size_t backupDim = g.outputDim;
 	g.Y = embedding;
 	g.outputDim = dimensions;
 
@@ -100,13 +100,13 @@ float* NNPNet::NNPNET::createPMDSEmbedding(Graph<float>& g, int& dimensions)
 		PivotMDS<double> pmds;
 		pmds.setNumberOfPivots(pmdsPivots);
 		pmds.call(g_d);
-		for (int i = 0; i < g.nodeCount * dimensions; i++) {
+		for (size_t i = 0; i < ((size_t)g.nodeCount) * ((size_t)dimensions); i++) {
 			g.Y[i] = (float)g_d.Y[i];
 		}
 	}
 
 	bool hasNaN = false;
-	for (int i = 0; i < g.nodeCount * dimensions; i++) {
+	for (size_t i = 0; i < ((size_t)g.nodeCount) * ((size_t)dimensions); i++) {
 		if (std::isnan(embedding[i])) {
 			hasNaN = true;
 			break;
@@ -494,13 +494,13 @@ static float* _smallEmbedding = nullptr;
 static bool first = true;
 
 PYBIND11_EMBEDDED_MODULE(getLists, m) {
-	m.def("getSmallEmbedding", [](int smallEmbeddingSize, int embeddingDim) {
+	m.def("getSmallEmbedding", [](size_t smallEmbeddingSize, size_t embeddingDim) {
 		return py::array_t<float>({ smallEmbeddingSize, embeddingDim }, _smallEmbedding);
 		});
-	m.def("getGt", [](int smallEmbeddingSize, int outputDim) {
+	m.def("getGt", [](size_t smallEmbeddingSize, size_t outputDim) {
 		return py::array_t<float>({ smallEmbeddingSize, outputDim }, _gt);
 		});
-	m.def("getFullEmbedding", [](int fullEmbeddingSize, int embeddingDim) {
+	m.def("getFullEmbedding", [](size_t fullEmbeddingSize, size_t embeddingDim) {
 		return py::array_t<float>({ fullEmbeddingSize, embeddingDim }, _fullEmbedding);
 		});
 }
